+++ CriticalMass/ManageFriendsViewController.swift
@@ -48,8 +48,10 @@ class ManageFriendsViewController: UIViewController, IBConstructable, UITableVie
         tableView.register(viewType: SettingsTableSectionHeader.self)
 
         //TODO: Fix collision with tableview gesture recognizer
-//        let tapGestureRecognizer = UITapGestureRecognizer(target: self, action: #selector(handleTap))
-//        view.addGestureRecognizer(tapGestureRecognizer)
+        let tapGestureRecognizer = UITapGestureRecognizer(target: self, action: #selector(handleTap(_:)))
+        tapGestureRecognizer.cancelsTouchesInView = false
+        view.addGestureRecognizer(tapGestureRecognizer)
+        
     }
 
     private func configureNavigationBar() {
@@ -67,7 +69,7 @@ class ManageFriendsViewController: UIViewController, IBConstructable, UITableVie
         navigationController?.pushViewController(viewController, animated: true)
     }
     
-    @objc private func handleTap() {
+    @objc private func handleTap(_ tap: UITapGestureRecognizer) {
         view.endEditing(true)
     }